\name{kissDE-package}
\alias{kissDE-package}
\alias{kissDE}
\docType{package}
\title{Differentially expressed events in RNA-seq data}
\description{This package retrieves differentially expressed events in RNA-seq data.
}
\details{
\tabular{ll}{
Package: \tab kissDE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-03\cr
License: \tab ceCill\cr
}
}

\author{
Lilia Brinza, Vincent Lacroix, Vincent Miele and Camille Marchet
Maintainers: vincent.miele@univ-lyon1.fr, camille.marchet@inria.fr
}
\keyword{ package }

